Extract BlogPostViewer into its own component
